Commands:
$ git fetch origin
$ git checkout -B ui-stable -t origin/ui-stable
$ git merge --strategy=ours origin/ui-canary
$ git diff --binary origin/ui-canary | git apply --reverse --index
$ git commit --amend

End state:
$ git diff ui-stable origin/ui-canary
[no output]
$ git rev-list --count ui-stable..origin/ui-canary
0

+++ src/trace_processor/util/proto_to_args_parser.h
@@ -143,7 +143,7 @@ class ProtoToArgsParser {
   // TODO(b/145578432): Add support for byte fields.
   base::Status ParseMessage(const protozero::ConstBytes& cb,
                             const std::string& type,
-                            const std::vector<uint16_t>* allowed_fields,
+                            const std::vector<uint32_t>* allowed_fields,
                             Delegate& delegate,
                             int* unknown_extensions = nullptr);
 
@@ -271,7 +271,7 @@ class ProtoToArgsParser {
   base::Status ParseMessageInternal(ScopedNestedKeyContext& key,
                                     const protozero::ConstBytes& cb,
                                     const std::string& type,
-                                    const std::vector<uint16_t>* fields,
+                                    const std::vector<uint32_t>* fields,
                                     Delegate& delegate,
                                     int* unknown_extensions);
